moves ScanTaskCommand and related method impls to its own module
